+++ lib/tasks/import.rake
@@ -23,7 +23,7 @@ namespace :import do
   task :todos_for_taxonomy_project, [:name, :csv_url] => :environment do |_, args|
     csv = RemoteCsv.new(args.csv_url)
     project = TaxonomyProject.create(args.name)
-    importer = Importers::TodosForTaxonomyProject.new(project, csv)
+    importer = TermGeneration::Importers::TodosForTaxonomyProject.new(project, csv)
     importer.run
     puts "Imported #{importer.completed.size} with #{importer.errors.size} errors"
   end
